refactor: move classes to domain package
