+++ backend/src/main/java/org/acme/TimetableConstraintProvider.java
@@ -5,6 +5,8 @@ import ai.timefold.solver.core.api.score.stream.Constraint;
 import ai.timefold.solver.core.api.score.stream.ConstraintFactory;
 import ai.timefold.solver.core.api.score.stream.ConstraintProvider;
 import ai.timefold.solver.core.api.score.stream.Joiners;
+import org.acme.domain.ConflictingUnit;
+import org.acme.domain.Unit;
 
 import java.util.function.Function;
