+++ backend/src/main/java/org/acme/domain/Unit.java
@@ -1,4 +1,4 @@
-package org.acme;
+package org.acme.domain;
 
 import ai.timefold.solver.core.api.domain.entity.PlanningEntity;
 import ai.timefold.solver.core.api.domain.lookup.PlanningId;
